# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: MentorLink API (minimal)
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /profile/me:
    get:
      summary: Get current profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile data
    put:
      summary: Update profile (including format preference)
      security:
        - bearerAuth: []
  /profile/competences:
    post:
      summary: Add competence
      security:
        - bearerAuth: []
  /profile/competences/{matiere_id}/activate:
    put:
      summary: Activate competence availability to help
      security:
        - bearerAuth: []
      parameters:
        - name: matiere_id
          in: path
          required: true
          schema:
            type: integer
  /profile/competences/{matiere_id}/deactivate:
    put:
      summary: Deactivate competence availability to help
      security:
        - bearerAuth: []
      parameters:
        - name: matiere_id
          in: path
          required: true
          schema:
            type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT